docmaster-backend/requirements.txt
fastapi
uvicorn[standard]
pymupdf4llm>=0.2.0
pdfplumber>=0.11.0
python-pptx>=0.6.0
python-multipart

# Optional: OCR fallback for empty PDF pages (pip install pytesseract Pillow; Tesseract 설치 필요)
# pytesseract
# Pillow
